Move modules to /src/server/app
