+++ src/server/app/orders/orders.module.ts
@@ -1,7 +1,7 @@
 import { forwardRef, Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { EmployeesModule } from 'src/server/employees/employees.module';
-import { VouchersModule } from 'src/server/vouchers/vouchers.module';
+import { EmployeesModule } from 'src/server/app/employees/employees.module';
+import { VouchersModule } from 'src/server/app/vouchers/vouchers.module';
 import { Order } from './order.entity';
 import { OrdersResolver } from './orders.resolver';
 import { OrdersService } from './orders.service';
