+++ src/server/app/partners/partners.resolver.ts
@@ -2,7 +2,7 @@ import { Resolver, Query, ResolveField, Parent } from '@nestjs/graphql';
 import { Inject } from '@nestjs/common';
 import { Partner } from './partner.entity';
 import { PartnersService } from './partners.service';
-import { VouchersService } from 'src/server/vouchers/vouchers.service';
+import { VouchersService } from 'src/server/app/vouchers/vouchers.service';
 
 @Resolver((_of) => Partner)
 export class PartnersResolver {
