+++ src/server/main.ts
@@ -1,5 +1,5 @@
 import { NestFactory } from '@nestjs/core';
-import { ServerModule } from './server.module';
+import { ServerModule } from 'src/server/server.module';
 
 async function bootstrap() {
   const app = await NestFactory.create(ServerModule);
