+++ src/server/server.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 
-import { AppModule } from './app.module';
+import { AppModule } from 'src/server/app/app.module';
 import { ViewModule } from 'src/server/view/view.module';
 
 @Module({
