+++ test/factories/employee.ts
@@ -1,6 +1,6 @@
 import { Factory } from 'fishery';
 import * as faker from 'faker';
-import { Employee } from 'src/server/employees/employee.entity';
+import { Employee } from 'src/server/app/employees/employee.entity';
 import companiesFactory from 'test/factories/company';
 
 export default Factory.define<Employee>(({ sequence }) => ({
