+++ test/factories/partner.ts
@@ -1,5 +1,5 @@
 import { Factory } from 'fishery';
-import { Partner } from 'src/server/partners/partner.entity';
+import { Partner } from 'src/server/app/partners/partner.entity';
 import * as faker from 'faker';
 
 export default Factory.define<Partner>(({ sequence }) => ({
